refactor: 필드 주입 방식 생성자 주입으로 변경
