+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,7 +1,6 @@
 package roomescape.service;
 
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
@@ -13,11 +12,13 @@ import roomescape.dto.response.ReservationResponse;
 @Service
 public class ReservationService {
 
-    @Autowired
-    ReservationDao reservationDao;
+    private final ReservationDao reservationDao;
+    private final ReservationTimeDao reservationTimeDao;
 
-    @Autowired
-    private ReservationTimeDao reservationTimeDao;
+    public ReservationService(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
+        this.reservationTimeDao = reservationTimeDao;
+    }
 
     public List<ReservationResponse> findAll() {
         List<Reservation> reservationDaoAll = reservationDao.findAll();
