+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,28 +1,24 @@
-package roomescape.controller;
+package roomescape.service;
 
 import jakarta.validation.Valid;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.request.ReservationTimeRequest;
 import roomescape.dto.response.ReservationTimeResponse;
 
-@RestController
-@RequestMapping(value = "/times")
-public class TimeController {
+@Service
+public class ReservationTimeService {
 
-    @Autowired
-    private ReservationTimeDao reservationTimeDao;
+    private final ReservationTimeDao reservationTimeDao;
 
-    @GetMapping()
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
+    }
+
     public List<ReservationTimeResponse> findAll() {
         List<ReservationTime> reservationTimeDaoAll = reservationTimeDao.findAll();
 
@@ -33,13 +29,11 @@ public class TimeController {
                 .toList();
     }
 
-    @PostMapping()
     public Long create(@Valid @RequestBody ReservationTimeRequest reservationTimeRequest) {
         ReservationTime reservationTime = reservationTimeRequest.toTime();
         return reservationTimeDao.create(reservationTime);
     }
 
-    @DeleteMapping("/{id}")
     public int delete(@PathVariable Long id) {
         return reservationTimeDao.delete(id);
     }
